fix: fix processFile usecase input

+++ internal/usecase/process_file.go
@@ -33,7 +33,7 @@ func NewProcessFile(
 }
 
 type ProcessFileInput struct {
-	FileID int
+	FileID int64
 	Path   string
 }
